+++ c++/assemble.sh
@@ -1,0 +1,2 @@
+cd ../data
+echo 1 | ../c++/bin/Linux/main false $1 0