+++ c++/run-debug.sh
@@ -1,0 +1,2 @@
+cd bin/Linux
+echo 2 | ./main true prog/verifCPU/verifCPU 0