+++ c++/run-test.sh
@@ -1,3 +1,2 @@
 cd bin/Linux
 echo 2 | ./main true prog/verifCPU/verifCPU 0 true
-cd ../..