+++ c++/run.sh
@@ -1,0 +1,2 @@
+cd bin/Linux
+echo 2 | ./main false $1 0